requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
databases==0.9.0
Django==5.1.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
dnspython==2.7.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scipy==1.14.1
setuptools==75.5.0
sqlparse==0.5.1
starlette==0.41.2
uvicorn==0.32.0